refactor(service): unify some agent services
