+++ apps/service/src/agents/service.ts
@@ -1,1100 +1,35 @@
-import { getHookByToken, getRun } from "workflow/api";
-
-import { loadKindConfig } from "@chia/agent-host/config";
-import {
-  AGENT_DELTA_NAMESPACE,
-  AGENT_TURN_KEY,
-  readAgentTurnMarker,
-} from "@chia/agent-host/execution";
-import type {
-  AgentStreamPosition,
-  AgentTurnMarker,
-} from "@chia/agent-host/execution";
 import type { AgentKindDefinition } from "@chia/agent-host/kind";
 import {
-  assertBelowRunningTurnCap,
-  assertWithinAgentQuota,
-} from "@chia/agent-host/quota";
-import { AGENT_TASK_IDS, resolveAgentTask } from "@chia/agent-host/tasks";
-import { recordAgentUsage } from "@chia/agent-host/usage";
-import {
   BYOK_PROVIDER_IDS,
-  createAgentModels,
   UnknownAgentModelError,
 } from "@chia/agent-runtime/models";
-import { canCompactBranch } from "@chia/agent-runtime/pi/compaction";
-import {
-  compactPiSession,
-  navigatePiSession,
-} from "@chia/agent-runtime/pi/maintenance";
-import {
-  computeSessionStats,
-  entriesUpToSeq,
-} from "@chia/agent-runtime/session/entries";
-import type { SessionEntry } from "@chia/agent-runtime/session/entries";
-import {
-  PgSessionRepo,
-  writeSessionSettings,
-} from "@chia/agent-runtime/session/pg-repo";
-import { walkBranch, walkTranscript } from "@chia/agent-runtime/session/tree";
-import type { SessionTree } from "@chia/agent-runtime/session/tree";
-import { estimateBranchContextTokens } from "@chia/agent-runtime/session/usage";
-import type {
-  AgentNavigationOptions,
-  AgentSessionDefaults,
-  AgentSessionSettings,
-  AgentUsageListener,
-  ThinkingLevel,
-  ToolTier,
-} from "@chia/agent-runtime/types";
-import { entriesToWireEvents } from "@chia/agent-runtime/wire/replay";
-import type { AgentWireEvent } from "@chia/agent-runtime/wire/schema";
-import type {
-  AgentKindService,
-  AgentServiceCaller,
-  AgentStreamCursor,
-} from "@chia/api/orpc/services/agent.service";
-import type { DB } from "@chia/db/client";
-import {
-  bindAgentRunExternalId,
-  completeAgentRun,
-  createAgentRun,
-  decideAgentApproval,
-  deleteAgentSession,
-  getActiveAgentRun,
-  getAgentApprovals,
-  getAgentSession,
-  getAgentSessionLastSeq,
-  patchAgentRunMetadata,
-  softDeleteAgentSession,
-  withAgentSessionLock,
-} from "@chia/db/repos/agent";
-import { AppError } from "@chia/service-kit/errors";
-import {
-  AGENT_END_SENTINEL,
-  agentMessageToken,
-} from "@chia/workflow-control/agent-hooks";
+import type { AgentKindService } from "@chia/api/orpc/services/agent.service";
 
-import { workflowControl } from "../repos/workflow-control.repo";
-import {
-  AGENT_ABORT_CONTROLLER_KEY,
-  readAgentAbortControllerRef,
-  signalAgentAbort,
-  startAgentAbortController,
-} from "../services/agent-abort-controller.service";
-import {
-  decryptAgentCredentials,
-  readEncryptedAgentCredentials,
-} from "../services/agent-credentials.service";
-import {
-  isRunLease,
-  isRunLive,
-  reconcileRunningAgentTurns,
-  runStateOf,
-} from "../services/agent-run-liveness.service";
+import { readEncryptedAgentCredentials } from "../services/agent-credentials.service";
+
+import { createAgentMaintenanceOperations } from "./service/maintenance";
+import { createAgentSessionOperations } from "./service/session";
+import { createAgentTurnOperations } from "./service/turn";
 
 /**
- * The `AgentKindService` for one registered kind.
+ * Builds the transport-facing service shared by every registered agent kind.
  *
- * Pi execution, session persistence, approval and wire primitives are `@chia/agent-runtime`; the
- * kind's tools, prompts and state are its {@link AgentKindDefinition}. This module is the
- * transport-facing glue between them and the durable workflow run, and it is the same for every
- * kind — nothing here knows what the kind does with a turn.
- *
- * **Stateless.** There is no in-process registry: each session is driven by a durable workflow run,
- * and every piece of state lives somewhere durable —
- *
- * - transcript → `agent.session_entry` (queried directly by the dashboard, and the source pi
- *   rebuilds context from)
- * - kind state → the kind's own tables, behind `definition.state`
- * - approval decisions → `agent.tool_approval`
- * - turn execution metadata → `agent.run`; pauses and event stream → the workflow backend
- *
- * That split is why a deploy mid-turn is survivable and why an approval can be granted a day later.
- * API instances can replicate because queue mutations cross `WorkflowControl` to the single
- * workflow process; no mutable conversation state lives in this process.
+ * Session persistence and projection, durable turn control, and tree maintenance are separate
+ * operation groups. This module composes them with the kind's model policy and capabilities; it
+ * owns no mutable process state.
  */
 export const createAgentKindService = <TState, TConfig extends object>(
   definition: AgentKindDefinition<TState, TConfig>
 ): AgentKindService => {
-  // ============================================
-  // Helpers
-  // ============================================
-  /** `defaults` only matter to `create`; the code's values are fine for every other operation. */
-  const repoFor = (
-    db: DB,
-    defaults: AgentSessionDefaults = definition.defaults
-  ) => new PgSessionRepo(db, { kind: definition.kind, defaults });
+  const sessions = createAgentSessionOperations(definition);
 
-  /**
-   * Loads a session row and its kind state **scoped to the caller**.
-   *
-   * The session id arrives from client input, so ownership is re-checked here rather than trusted —
-   * the guard proves who is calling, not what they may open.
-   */
-  const loadOwnedRow = async (
-    caller: AgentServiceCaller,
-    sessionId: string
-  ) => {
-    const db = caller.context.db;
-    const row = await getAgentSession(db, sessionId);
-    if (!row || row.deletedAt !== null) return null;
-    if (row.userId !== caller.userId) return null;
-    if (row.kind !== definition.kind) return null;
-
-    const state = await definition.state.load(db, sessionId);
-    if (state === null) return null;
-
-    return { ...row, state };
-  };
-
-  type OwnedRow = NonNullable<Awaited<ReturnType<typeof loadOwnedRow>>>;
-
-  /** {@link loadOwnedRow} plus the session's active durable run, for everything that acts on it. */
-  const loadOwnedSession = async (
-    caller: AgentServiceCaller,
-    sessionId: string
-  ) => {
-    const row = await loadOwnedRow(caller, sessionId);
-    if (!row) return null;
-
-    const activeRun = await getActiveAgentRun(caller.context.db, sessionId);
-    return {
-      ...row,
-      activeRunId: activeRun?.id ?? null,
-      workflowRunId: activeRun?.externalRunId ?? null,
-      startedAt: activeRun?.startedAt ?? null,
-      turn: activeRun ? readAgentTurnMarker(activeRun.metadata) : undefined,
-      abortController: activeRun
-        ? readAgentAbortControllerRef(activeRun.metadata)
-        : undefined,
-    };
-  };
-
-  type OwnedSession = NonNullable<Awaited<ReturnType<typeof loadOwnedSession>>>;
-
-  /**
-   * The caller with its database handle swapped for the lock's transaction, so everything an
-   * operation does under `withAgentSessionLock` runs on the lock's own connection.
-   */
-  const withDb = (caller: AgentServiceCaller, db: DB): AgentServiceCaller => ({
-    ...caller,
-    context: { ...caller.context, db },
-  });
-
-  const settingsOf = (row: {
-    id: string;
-    providerId: string | null;
-    modelId: string | null;
-    thinkingLevel: string | null;
-    activeToolNames: string[] | null;
-    autoApprove: string[];
-  }): AgentSessionSettings => {
-    if (!row.providerId || !row.modelId || !row.thinkingLevel) {
-      throw new Error(`Agent session ${row.id} has incomplete LLM settings.`);
-    }
-    return {
-      providerId: row.providerId,
-      modelId: row.modelId,
-      thinkingLevel:
-        /* SAFETY: The producer contract guarantees this value satisfies ThinkingLevel. */ row.thinkingLevel as ThinkingLevel,
-      activeToolNames: row.activeToolNames,
-      autoApprove:
-        /* SAFETY: The producer contract guarantees this value satisfies ToolTier[]. */ row.autoApprove as ToolTier[],
-    };
-  };
-
-  const summaryOf = (row: OwnedRow) => {
-    const settings = settingsOf(row);
-    return {
-      id: row.id,
-      title: row.title,
-      kind: row.kind,
-      modelId: settings.modelId,
-      thinkingLevel: settings.thinkingLevel,
-      ...definition.state.summary(row.state),
-      forkedFromSessionId: row.forkedFromSessionId,
-      createdAt: row.createdAt.getTime(),
-      updatedAt: row.updatedAt.getTime(),
-    };
-  };
-
-  /** Per-request `Models`, carrying whatever provider keys the caller has registered. */
-  const modelsFor = (caller: AgentServiceCaller) =>
-    createAgentModels(
-      decryptAgentCredentials(
-        readEncryptedAgentCredentials(caller.context.headers)
-      )
-    );
-
-  /**
-   * A maintenance operation carries one model call — the summary — inside the request, under
-   * the session lock and its transaction. The RPC route exempts it from the shared request
-   * timeout (a 504 there only drops the response while the work and the lock run on), so this
-   * deadline is what bounds the lock: the summary is cancelled, nothing is appended, and the
-   * transaction rolls back with the error.
-   */
-  const MAINTENANCE_DEADLINE_MS = 120_000;
-  const maintenanceTimedOut = (action: string) =>
-    new AppError("TIMEOUT", {
-      message: `Could not ${action} within ${MAINTENANCE_DEADLINE_MS / 1000}s. The conversation is unchanged.`,
-    });
-  const nothingToCompact = () =>
-    new AppError("CONFLICT", {
-      message:
-        "Nothing to compact: the conversation still fits in what a compaction keeps.",
-    });
-
-  /**
-   * Maintenance operations over the session tree.
-   *
-   * These only walk the tree, so no tools, ports, approval gate or event subscriptions are
-   * built — just the session, its settings and a model. Which model is the compaction or
-   * branch-summary *task*'s to say: by default the session's own, which a BYOK session needs
-   * its key for here just as much as in a turn (this path runs inside the request, so the
-   * cookie is read and decrypted directly rather than travelling through the workflow); pinned
-   * by the operator, a house model that needs no key at all. The session model is resolved
-   * only when the task follows it.
-   *
-   * `signal` bounds the one model call — the summary — and nothing else: when it fires the
-   * summary is cancelled and the tree is untouched.
-   */
-  const maintenanceFor = async (
-    caller: AgentServiceCaller,
-    row: OwnedRow,
-    signal: AbortSignal
-  ) => {
-    const db = caller.context.db;
-    const session = await repoFor(db).openById(row.id);
-    const settings = settingsOf(row);
-    const models = modelsFor(caller);
-    // Written on the lock's transaction, so the row lands with the entry it accounts for.
-    const onUsage: AgentUsageListener = (report) =>
-      recordAgentUsage(db, {
-        userId: caller.userId,
-        sessionId: row.id,
-        kind: definition.kind,
-        ...report,
-      });
-    const operationFor = async (taskId: string) => {
-      const task = await resolveAgentTask(db, taskId, {
-        session: () => ({
-          model: definition.models.resolve(settings, models),
-          models,
-        }),
-      });
-      return {
-        session,
-        settings,
-        model: task.model,
-        models: task.models,
-        signal,
-        onUsage,
-      };
-    };
-    return {
-      session,
-      compact: async (customInstructions?: string) =>
-        compactPiSession(
-          await operationFor(AGENT_TASK_IDS.sessionCompaction),
-          customInstructions
-        ),
-      navigate: async (entryId: string, options: AgentNavigationOptions) =>
-        navigatePiSession(
-          await operationFor(AGENT_TASK_IDS.sessionBranchSummary),
-          entryId,
-          options
-        ),
-    };
-  };
-
-  const replayOptions = {
-    tierOf: definition.policy.tierOf,
-    labelOf: definition.policy.labelOf,
-    summarize: definition.policy.summarize,
-  };
-
-  /**
-   * A promise that never settles, used to drop an exhausted reader out of the two-stream race in
-   * `stream` without it winning again.
-   */
-  const NEVER = new Promise<never>(() => undefined);
-
-  /**
-   * Marks the run's next turn as running before the workflow is woken to run it, so maintenance
-   * refuses from the moment a turn is accepted rather than from the moment its step starts. The
-   * step rewrites the marker with the same leaf and index when it begins. A turn already running
-   * keeps its marker: this one queues behind it, and the step marks it when its own turn comes.
-   */
-  const claimTurn = async (
-    db: DB,
-    row: OwnedSession,
-    position: AgentStreamPosition
-  ): Promise<void> => {
-    if (!row.activeRunId || row.turn?.running) return;
-    const turn: AgentTurnMarker = {
-      seqBefore: await getAgentSessionLastSeq(db, row.id),
-      ...position,
-      running: true,
-    };
-    await patchAgentRunMetadata(db, row.activeRunId, {
-      [AGENT_TURN_KEY]: turn,
-    });
-  };
-
-  /**
-   * Where the next turn on a live run will begin: one past the current tail of each durable
-   * stream. Captured before the run is woken, so a turn still active shares this continuation
-   * with the queued one in durable emission order.
-   */
-  const nextTurnPosition = async (
-    runId: string
-  ): Promise<AgentStreamPosition> => {
-    const run = getRun(runId);
-    const [coarseTail, deltaTail] = await Promise.all([
-      run.getReadable().getTailIndex(),
-      run.getReadable({ namespace: AGENT_DELTA_NAMESPACE }).getTailIndex(),
-    ]);
-    return { streamIndex: coarseTail + 1, deltaStreamIndex: deltaTail + 1 };
-  };
-
-  const cursorOf = (
-    runId: string,
-    position: AgentStreamPosition
-  ): AgentStreamCursor => ({
-    runId,
-    startIndex: position.streamIndex,
-    deltaStartIndex: position.deltaStreamIndex,
-  });
-
-  /**
-   * How long `abort` waits for the stopped turn's `run:end` before cancelling the run regardless.
-   * A tool that ignores its signal keeps Pi waiting on it; the cancel then proceeds and the
-   * tool's result, if it ever comes, is closed on replay as aborted.
-   */
-  const ABORT_SETTLE_TIMEOUT_MS = 10_000;
-
-  /**
-   * Waits for the turn step to write the stopped turn's `run:end`.
-   *
-   * The signal stops Pi, but the step still has the partial reply and any in-flight tool result
-   * to persist before the terminal event; the client rebuilds the transcript the moment `abort`
-   * returns, so returning earlier would show it a turn still missing its last entries. The turn's
-   * own durable stream is the wait: `runPiTurn` appends every entry before emitting its wire
-   * event, so `run:end` arriving means the transcript is complete.
-   */
-  const waitForTurnEnd = async (
-    runId: string,
-    startIndex: number
-  ): Promise<void> => {
-    const reader = getRun(runId)
-      .getReadable<AgentWireEvent>({ startIndex })
-      .getReader();
-    // Cancelling settles the pending read as done; the loop then falls through.
-    const deadline = setTimeout(
-      () => void reader.cancel().catch(() => undefined),
-      ABORT_SETTLE_TIMEOUT_MS
-    );
-    try {
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done || value?.type === "run:end") return;
-      }
-    } catch {
-      // A dropped stream has nothing more to tell; the cancel below proceeds as before.
-    } finally {
-      clearTimeout(deadline);
-      await reader.cancel().catch(() => undefined);
-    }
-  };
-
-  /**
-   * Cancels a run that was live a moment ago.
-   *
-   * `cancel()` refuses a run that has since reached a terminal state — the storage layer
-   * throws `EntityConflictError` rather than no-op — so the check is redone on failure: a
-   * run that finished on its own is exactly the outcome the caller wanted, while a run
-   * that is still live means the cancel genuinely failed and must surface.
-   */
-  const cancelLiveRun = async (runId: string): Promise<void> => {
-    try {
-      await workflowControl.cancelRun(runId);
-    } catch (error) {
-      if (await isRunLive(runId)) throw error;
-    }
-  };
-
-  /**
-   * Whether a hook token is registered and can be resumed.
-   *
-   * `createHook()` does not register on call. This workflow registers through `getConflict()` before
-   * its first turn, but `start()` may return before the workflow reaches that line. Checking once
-   * turns that startup race into a retryable answer instead of a failed `resume()`; it is not a
-   * timer or polling loop.
-   */
-  const isHookReady = async (token: string): Promise<boolean> => {
-    try {
-      return Boolean(await getHookByToken(token));
-    } catch {
-      return false;
-    }
-  };
-
-  /**
-   * Approvals still awaiting a decision.
-   *
-   * Used to refuse a new prompt while the run is parked on an approval hook. The message would
-   * otherwise sit in the event log unread until the approval resolved, which looks to the operator
-   * like their message vanished.
-   */
-  const undecidedApprovals = async (
-    db: DB,
-    sessionId: string
-  ): Promise<string[]> => {
-    const approvals = await getAgentApprovals(db, sessionId);
-    return approvals
-      .filter((approval) => approval.decidedAt === null)
-      .map((approval) => approval.toolName);
-  };
-
-  /**
-   * The branch as it was persisted before the running turn started. Everything the turn appends
-   * — its own user message included — is replayed by `attach` from the run's stream, which starts
-   * at the marker's `streamIndex`, before the turn announces `user`. Taking the user message from
-   * both sources showed it twice to a client that rejoined mid-turn.
-   */
-  const entriesBeforeTurn = (
-    entries: SessionEntry[],
-    turn: AgentTurnMarker
-  ): SessionEntry[] => entriesUpToSeq(entries, turn.seqBefore);
-
-  const detailFor = async (caller: AgentServiceCaller, sessionId: string) => {
-    const row = await loadOwnedSession(caller, sessionId);
-    if (!row) return null;
-
-    const db = caller.context.db;
-    const session = await repoFor(db).openById(sessionId);
-
-    // One query at a time: under the session lock `db` is a transaction on a single connection,
-    // and pg queues — and deprecates — a second query issued on a client that is busy.
-    const entries = await session.getEntries();
-    const leafId = await session.getLeafId();
-    /**
-     * Two walks of one load. The branch stops at the newest compaction and is what the model's
-     * next request carries — the context estimate and `compactable` read it. The transcript
-     * runs through every compaction: what was said stays on screen, with the compaction shown
-     * where it happened, because it condensed the model's context, not the conversation.
-     */
-    const branch = walkBranch(entries, leafId);
-    const transcript = walkTranscript(entries, leafId);
-    const stats = computeSessionStats(entries);
-    const kindDetail = await definition.state.detail(db, sessionId, row.state);
-    const approvals = await getAgentApprovals(db, sessionId);
-    /**
-     * Read after the branch, on purpose: the cut below and the entries it cuts must come from the same
-     * observation. A turn that finished between the two reads has already persisted its reply; a
-     * marker read before the branch would still say `running`, cut that reply out, and leave nothing
-     * for `attach` to replay.
-     */
-    const activeRun = await getActiveAgentRun(db, sessionId);
-    const turn = activeRun
-      ? readAgentTurnMarker(activeRun.metadata)
-      : undefined;
-    const run = await runStateOf({
-      activeRunId: activeRun?.id ?? null,
-      workflowRunId: activeRun?.externalRunId ?? null,
-      startedAt: activeRun?.startedAt ?? null,
-      turn,
-    });
-
-    /**
-     * A running turn is not replayed from the transcript: `attach` replays it from the run's stream
-     * instead, and both cut at the same recorded marker so the client sees each message once.
-     */
-    const transcriptEntries =
-      run?.status === "running" && turn
-        ? entriesBeforeTurn(transcript, turn)
-        : transcript;
-    const contextEntries =
-      run?.status === "running" && turn
-        ? entriesBeforeTurn(branch, turn)
-        : branch;
-    const events = entriesToWireEvents(transcriptEntries, replayOptions);
-
-    // Approval events are never replayed from the transcript; the rows are. A pending row restores
-    // the prompt on reload, a decided one closes its card the way the live stream did.
-    const approvalRows = approvals.map((approval) => ({
-      toolCallId: approval.toolCallId,
-      toolName: approval.toolName,
-      args: approval.args ?? undefined,
-      status: approval.status,
-      comment: approval.comment ?? undefined,
-    }));
-
-    return {
-      session: summaryOf(row),
-      settings: settingsOf(row),
-      runtimeConfig: row.runtimeConfig,
-      configVersion: row.configVersion,
-      ...kindDetail,
-      run,
-      events,
-      approvals: approvalRows,
-      stats: {
-        messageCount: stats.messageCount,
-        contextTokens: estimateBranchContextTokens(contextEntries),
-        compactable: canCompactBranch(contextEntries),
-        totalTokens: stats.totalTokens,
-        costTotal: stats.costTotal,
-      },
-    };
-  };
-
-  /**
-   * Maintenance (compact, rewind, fork) mutates or copies the session tree, so it cannot run
-   * while a turn is appending to it. A live run is not enough to refuse on — parked on the
-   * message hook is its normal idle state — so this reads the turn marker, like everything else
-   * that needs to know. Called under the session lock, after the row was read under it: a turn
-   * accepted before the lock was taken is already marked, and one accepted after waits.
-   *
-   * An undecided approval refuses too: the run is parked on the approval hook, and the relay
-   * turn its decision starts would land on whatever branch is active then — answering a call
-   * that is no longer on it. `CONFLICT`, converted at the route.
-   */
-  const assertMaintainable = async (
-    row: OwnedSession,
-    db: DB,
-    action: string
-  ): Promise<void> => {
-    if ((await runStateOf(row))?.status === "running") {
-      throw new AppError("CONFLICT", {
-        message: `Cannot ${action} while a turn is running. Wait for it to finish or abort it.`,
-      });
-    }
-    const outstanding = await undecidedApprovals(db, row.id);
-    if (outstanding.length > 0) {
-      throw new AppError("CONFLICT", {
-        message: `Cannot ${action} while \`${outstanding.join("`, `")}\` awaits your decision. Approve or reject it first.`,
-      });
-    }
-  };
-
-  /** The entry a rewind or fork targets, or `NOT_FOUND`; the client only ever holds ids it was shown. */
-  const requireEntry = async (
-    session: SessionTree,
-    entryId: string
-  ): Promise<SessionEntry> => {
-    const entry = await session.getEntry(entryId);
-    if (!entry) {
-      throw new AppError("NOT_FOUND", {
-        message: `Entry ${entryId} is not in this session.`,
-      });
-    }
-    return entry;
-  };
-
-  // ============================================
-  // Service
-  // ============================================
-
-  const service: AgentKindService = {
+  return {
     minTier: definition.minTier,
+    ...sessions.service,
+    ...createAgentTurnOperations(definition, sessions),
+    ...createAgentMaintenanceOperations(definition, sessions),
 
-    async listSessions(caller, input) {
-      const db = caller.context.db;
-      const metadata = await repoFor(db).list({
-        userId: caller.userId,
-        limit: input?.limit,
-        includeDeleted: input?.includeDeleted,
-      });
-
-      const rows = await Promise.all(
-        metadata.map((entry) => loadOwnedRow(caller, entry.id))
-      );
-
-      return {
-        items: rows.flatMap((row) => (row ? [summaryOf(row)] : [])),
-        // Sessions are listed newest-first with a hard limit; there is no cursor to page yet.
-        nextCursor: null,
-      };
-    },
-
-    async createSession(caller, input) {
-      const db = caller.context.db;
-      // The operator's defaults for this kind, read now: a session copies them onto its own row.
-      const { defaults } = await loadKindConfig(db, definition);
-      const session = await repoFor(db, defaults).create({
-        userId: caller.userId,
-        title: input.title,
-        settings: {
-          providerId: input.model?.providerId,
-          modelId: input.model?.modelId,
-          thinkingLevel:
-            /* SAFETY: The producer contract guarantees this value satisfies ThinkingLevel | undefined. */ input.thinkingLevel as
-              | ThinkingLevel
-              | undefined,
-          autoApprove:
-            /* SAFETY: The producer contract guarantees this value satisfies ToolTier[] | undefined. */ input.autoApprove as
-              | ToolTier[]
-              | undefined,
-        },
-        runtimeConfig: input.runtimeConfig,
-      });
-      const { id } = session;
-      try {
-        await definition.state.create(caller, db, id, input);
-
-        const detail = await detailFor(caller, id);
-        if (!detail)
-          throw new Error("Session vanished immediately after creation");
-        return detail;
-      } catch (error) {
-        // Core and kind state live in separate tables. Compensate if the kind's setup fails so
-        // callers never receive an unusable half-created session.
-        await deleteAgentSession(db, id);
-        throw error;
-      }
-    },
-
-    getSession(caller, input) {
-      return detailFor(caller, input.sessionId);
-    },
-
-    async deleteSession(caller, input) {
-      const row = await loadOwnedSession(caller, input.sessionId);
-      if (!row) return false;
-
-      // End the run before soft-deleting, so it is not left parked on a hook forever.
-      // Cancelled rather than sent the end sentinel: the run may be parked on an
-      // *approval* hook, where a queued message is never read — and once the session is
-      // deleted nobody can decide the approval, so the run would stay parked for good.
-      if (row.workflowRunId && (await isRunLive(row.workflowRunId))) {
-        await cancelLiveRun(row.workflowRunId);
-      }
-      if (row.activeRunId) {
-        await completeAgentRun(caller.context.db, row.activeRunId, "cancelled");
-      }
-
-      await softDeleteAgentSession(caller.context.db, input.sessionId);
-      return true;
-    },
-
-    async updateSettings(caller, input) {
-      const row = await loadOwnedSession(caller, input.sessionId);
-      if (!row) return null;
-      await writeSessionSettings(caller.context.db, input.sessionId, {
-        title: input.title,
-        providerId: input.model?.providerId,
-        modelId: input.model?.modelId,
-        thinkingLevel:
-          /* SAFETY: The producer contract guarantees this value satisfies ThinkingLevel | undefined. */ input.thinkingLevel as
-            | ThinkingLevel
-            | undefined,
-        activeToolNames: input.activeToolNames,
-        autoApprove:
-          /* SAFETY: The producer contract guarantees this value satisfies ToolTier[] | undefined. */ input.autoApprove as
-            | ToolTier[]
-            | undefined,
-        runtimeConfig: input.runtimeConfig,
-      });
-
-      return detailFor(caller, input.sessionId);
-    },
-
-    /**
-     * Enqueues a turn.
-     *
-     * If the session already has a live run, its reusable hook durably queues the message. The
-     * workflow consumes it after the current turn and any approval handshake finish. Otherwise a new
-     * run is started and its id recorded.
-     *
-     * Either way this returns as soon as the message is accepted; the turn itself runs in the run.
-     * The whole step holds the session lock, so maintenance can neither slip in between the turn
-     * being accepted and its marker being visible, nor be mid-mutation when the turn starts.
-     */
-    prompt: (outer, input) =>
-      withAgentSessionLock(outer.context.db, input.sessionId, async (tx) => {
-        const caller = withDb(outer, tx);
-        const db = tx;
-        const row = await loadOwnedSession(caller, input.sessionId);
-        if (!row) throw new Error(`Unknown agent session: ${input.sessionId}`);
-
-        if (input.text === AGENT_END_SENTINEL) {
-          throw new Error(
-            `"${AGENT_END_SENTINEL}" is reserved; it ends the session's run.`
-          );
-        }
-
-        // Before anything is queued: a turn accepted here is one the house pays for. Checked under
-        // the lock so two prompts on one session cannot both pass on the same reading; the cap takes
-        // the user's lock too, against two prompts on two sessions. Dead runs are closed first, so a
-        // step the process died under cannot hold a slot of the cap.
-        await assertWithinAgentQuota(db, caller);
-        await reconcileRunningAgentTurns(db, caller.userId);
-        await assertBelowRunningTurnCap(db, caller);
-
-        const message = {
-          text: input.text,
-          template: input.template,
-          preAuthorizeToolNames: input.preAuthorizeToolNames,
-          // Refreshed on every prompt: the run outlives any one request, and the operator may have
-          // registered or rotated a key since the last turn.
-          credentials: readEncryptedAgentCredentials(caller.context.headers),
-        };
-
-        /**
-         * Refuse while an approval is outstanding.
-         *
-         * The run is parked on the *approval* hook, not the message hook. A resumed message would be
-         * persisted and then sit unread until the approval resolved — from the operator's side their
-         * message would simply appear to do nothing. Better to say why.
-         */
-        const outstanding = await undecidedApprovals(
-          caller.context.db,
-          input.sessionId
-        );
-        if (outstanding.length > 0) {
-          throw new Error(
-            `Waiting on your decision for \`${outstanding.join("`, `")}\`. Approve or reject it before sending another message.`
-          );
-        }
-
-        if (row.workflowRunId && (await isRunLive(row.workflowRunId))) {
-          const token = agentMessageToken(input.sessionId);
-
-          if (!(await isHookReady(token))) {
-            throw new Error(
-              "The session's run is still starting up. Retry in a moment."
-            );
-          }
-
-          const position = await nextTurnPosition(row.workflowRunId);
-          await claimTurn(db, row, position);
-          await workflowControl.resumeAgentMessage(input.sessionId, message);
-          return {
-            ...cursorOf(row.workflowRunId, position),
-            startedRun: false,
-          };
-        }
-
-        // Started before the session run so its ref can travel in the run's request: every turn then
-        // subscribes to this one controller by run id, and `abort` resumes it by id.
-        const abortController = await startAgentAbortController();
-
-        // The row is written before the workflow exists and is the session's turn lease from then on:
-        // maintenance that takes the lock after this already sees a running turn, and the step's marker
-        // writes (addressed by session) always find their row. The workflow backend mints the run id,
-        // so the row's own id stands in as `externalRunId` until the started run is bound to it; a
-        // start that fails closes the row so the lease does not outlive the attempt.
-        const runId = crypto.randomUUID();
-        const position: AgentStreamPosition = {
-          streamIndex: 0,
-          deltaStreamIndex: 0,
-        };
-        const turn: AgentTurnMarker = {
-          seqBefore: await getAgentSessionLastSeq(db, row.id),
-          ...position,
-          running: true,
-        };
-        await createAgentRun(db, {
-          id: runId,
-          sessionId: input.sessionId,
-          harnessKind: "workflow",
-          externalRunId: runId,
-          metadata: {
-            agentKind: definition.kind,
-            [AGENT_TURN_KEY]: turn,
-            [AGENT_ABORT_CONTROLLER_KEY]: {
-              id: abortController.id,
-              runId: abortController.runId,
-            },
-          },
-        });
-        let workflowRunId;
-        try {
-          workflowRunId = await workflowControl.startAgentSession({
-            sessionId: input.sessionId,
-            runId,
-            userId: caller.userId,
-            abortController,
-            firstMessage: message,
-          });
-        } catch (error) {
-          await completeAgentRun(db, runId, "failed");
-          throw error;
-        }
-        await bindAgentRunExternalId(db, runId, workflowRunId);
-
-        return { ...cursorOf(workflowRunId, position), startedRun: true };
-      }),
-
-    async attach(caller, input) {
-      const row = await loadOwnedSession(caller, input.sessionId);
-      if (!row?.workflowRunId || !row.turn) return null;
-      // A lease has no stream to tail yet; the client's next `get` finds the bound run.
-      if (isRunLease(row)) return null;
-      const run = await runStateOf(row);
-      if (run?.status !== "running") return null;
-      return cursorOf(row.workflowRunId, row.turn);
-    },
-
-    /**
-     * Tails a run's durable stream.
-     *
-     * The heavy lifting is the SDK's: `getReadable({ startIndex })` replays from the requested point
-     * and then delivers live chunks, which is what makes reconnection and multi-viewer work without
-     * any coordination on our side.
-     */
-    async *stream(caller, input) {
-      const row = await loadOwnedSession(caller, input.sessionId);
-      if (!row) throw new Error(`Unknown agent session: ${input.sessionId}`);
-
-      const runId = input.runId ?? row.workflowRunId;
-      if (!runId) return;
-
-      const readable = getRun(runId).getReadable<AgentWireEvent>({
-        startIndex: input.startIndex,
-      });
-      const reader = readable.getReader();
-
-      // Deltas live on their own namespace and arrive batched; merging them here keeps the client's
-      // reducer identical whether or not it asked for them.
-      const deltaReader =
-        input.deltaStartIndex === undefined
-          ? undefined
-          : getRun(runId)
-              .getReadable<AgentWireEvent[]>({
-                namespace: AGENT_DELTA_NAMESPACE,
-                startIndex: input.deltaStartIndex,
-              })
-              .getReader();
-
-      try {
-        if (!deltaReader) {
-          while (true) {
-            const { done, value } = await reader.read();
-            if (done) break;
-            if (value) yield value;
-          }
-          return;
-        }
-
-        /**
-         * Two streams, drained as whichever has data next.
-         *
-         * Racing the two reads (rather than draining one then the other) is what keeps deltas
-         * interleaved with the coarse events they belong to.
-         */
-        let coarsePending = reader.read();
-        let deltaPending = deltaReader.read();
-        let coarseDone = false;
-        let deltaDone = false;
-
-        while (!coarseDone || !deltaDone) {
-          const winner = await Promise.race([
-            coarsePending.then((result) => ({
-              kind: "coarse" as const,
-              result,
-            })),
-            deltaPending.then((result) => ({
-              kind: "delta" as const,
-              result,
-            })),
-          ]);
-
-          if (winner.kind === "coarse") {
-            if (winner.result.done) {
-              coarseDone = true;
-              // A never-settling promise keeps the exhausted side out of the race.
-              coarsePending = NEVER;
-            } else {
-              if (winner.result.value) yield winner.result.value;
-              coarsePending = reader.read();
-            }
-          } else {
-            if (winner.result.done) {
-              deltaDone = true;
-              deltaPending = NEVER;
-            } else {
-              for (const event of winner.result.value ?? []) yield event;
-              deltaPending = deltaReader.read();
-            }
-          }
-        }
-      } finally {
-        await reader.cancel().catch(() => undefined);
-        await deltaReader?.cancel().catch(() => undefined);
-      }
-    },
-
-    async abort(caller, input) {
-      const row = await loadOwnedSession(caller, input.sessionId);
-      if (!row?.workflowRunId) return false;
-      if (!(await isRunLive(row.workflowRunId))) return false;
-
-      // Stop the harness first — cancelling the run does not reach a step already in flight — then
-      // cancel the whole run, which is the session's driver; the next prompt starts a fresh one and
-      // picks the transcript back up from Postgres. The row is marked last so a failed cancel never
-      // leaves a live run behind a non-active row (the next prompt would start a second workflow and
-      // hit the hook conflict).
-      if (row.abortController) {
-        const signalled = await signalAgentAbort(
-          row.abortController.id,
-          "stopped by the operator"
-        );
-        if (signalled && row.turn?.running) {
-          await waitForTurnEnd(row.workflowRunId, row.turn.streamIndex);
-        }
-      }
-      await cancelLiveRun(row.workflowRunId);
-      if (row.activeRunId) {
-        await completeAgentRun(caller.context.db, row.activeRunId, "cancelled");
-      }
-      return true;
-    },
-
-    /** The decision starts a relay turn, so it is accepted under the session lock like a prompt. */
-    approve: (outer, input) =>
-      withAgentSessionLock(outer.context.db, input.sessionId, async (tx) => {
-        const caller = withDb(outer, tx);
-        const db = tx;
-        const row = await loadOwnedSession(caller, input.sessionId);
-        if (!row?.workflowRunId) return null;
-
-        // The decision starts a relay turn the model answers, so it is metered and capped like a
-        // prompt; refused before it is persisted, so the approval stays pending until it may run.
-        await assertWithinAgentQuota(db, caller);
-        await reconcileRunningAgentTurns(db, caller.userId);
-        await assertBelowRunningTurnCap(db, caller);
-
-        // Persist first: the decision must outlive the run, and the permission gate reads it back from
-        // here when the tool call is re-issued.
-        const decided = await decideAgentApproval(db, {
-          sessionId: input.sessionId,
-          toolCallId: input.toolCallId,
-          approved: input.approved,
-          comment: input.comment,
-          decidedBy: caller.userId,
-        });
-        if (!decided) return null;
-
-        // Capture the cursor before waking the workflow. The chat transport opens a fresh request for
-        // an approval continuation, so it needs the same exact replay boundary as a normal prompt.
-        const position = await nextTurnPosition(row.workflowRunId);
-        await claimTurn(db, row, position);
-
-        // Then wake the run, which has been parked on this hook with no compute consumed.
-        await workflowControl.resumeAgentApproval(
-          input.sessionId,
-          input.toolCallId,
-          {
-            approved: input.approved,
-            comment: input.comment,
-            // The turns the workflow synthesises after this decision have no request of their own.
-            credentials: readEncryptedAgentCredentials(caller.context.headers),
-          }
-        );
-
-        return cursorOf(row.workflowRunId, position);
-      }),
-
-    compact: (outer, input) =>
-      withAgentSessionLock(outer.context.db, input.sessionId, async (tx) => {
-        const caller = withDb(outer, tx);
-        const row = await loadOwnedSession(caller, input.sessionId);
-        if (!row) return null;
-        await assertMaintainable(row, tx, "compact");
-
-        const deadline = AbortSignal.timeout(MAINTENANCE_DEADLINE_MS);
-        const maintenance = await maintenanceFor(caller, row, deadline);
-        // Only a summary calls the model, so only a branch with something to condense needs quota.
-        if (!canCompactBranch(await maintenance.session.getBranch())) {
-          throw nothingToCompact();
-        }
-        await assertWithinAgentQuota(tx, caller);
-
-        let compacted: Awaited<ReturnType<typeof maintenance.compact>>;
-        try {
-          compacted = await maintenance.compact(input.customInstructions);
-        } catch (error) {
-          if (deadline.aborted) throw maintenanceTimedOut("compact");
-          throw error;
-        }
-        if (!compacted) throw nothingToCompact();
-        return detailFor(caller, input.sessionId);
-      }),
-
-    navigate: (outer, input) =>
-      withAgentSessionLock(outer.context.db, input.sessionId, async (tx) => {
-        const caller = withDb(outer, tx);
-        const row = await loadOwnedSession(caller, input.sessionId);
-        if (!row) return null;
-        await assertMaintainable(row, tx, "rewind");
-        // Only a summary calls the model; moving the leaf is free.
-        if (input.summarize) await assertWithinAgentQuota(tx, caller);
-
-        const deadline = AbortSignal.timeout(MAINTENANCE_DEADLINE_MS);
-        const maintenance = await maintenanceFor(caller, row, deadline);
-        await requireEntry(maintenance.session, input.entryId);
-        const result = await maintenance.navigate(input.entryId, {
-          summarize: input.summarize,
-          label: input.label,
-        });
-        if (result.cancelled) throw maintenanceTimedOut("rewind");
-        return detailFor(caller, input.sessionId);
-      }),
-
-    fork: (outer, input) =>
-      withAgentSessionLock(outer.context.db, input.sessionId, async (tx) => {
-        const caller = withDb(outer, tx);
-        const db = tx;
-        const row = await loadOwnedSession(caller, input.sessionId);
-        if (!row) return null;
-        await assertMaintainable(row, db, "fork");
-
-        const repo = repoFor(db);
-        const position = input.position ?? "before";
-        if (input.entryId) {
-          const target = await requireEntry(
-            await repo.openById(row.id),
-            input.entryId
-          );
-          if (
-            position === "before" &&
-            (target.type !== "message" || target.message.role !== "user")
-          ) {
-            throw new AppError("BAD_REQUEST", {
-              message:
-                "Only a user message can be forked before; fork at this entry instead.",
-            });
-          }
-        }
-
-        const forked = await repo.fork(
-          { id: row.id },
-          { entryId: input.entryId, position, title: input.title }
-        );
-        try {
-          await definition.state.fork(db, row.id, forked.id);
-          const detail = await detailFor(caller, forked.id);
-          if (!detail)
-            throw new Error("Session vanished immediately after fork");
-          return detail;
-        } catch (error) {
-          // Same compensation as `createSession`: a fork without its kind state can never be opened.
-          await deleteAgentSession(db, forked.id);
-          throw error;
-        }
-      }),
-
-    /**
-     * The pre-persistence check the transport calls.
-     *
-     * Returns a reason instead of throwing because the caller is a middleware turning this into a
-     * `BAD_REQUEST`. The kind's `assert` checks policy *and* catalogue membership — the latter is
-     * what stops a typo on a native provider from being stored and then failing on every subsequent
-     * turn, deep in the workflow step where the operator cannot see why.
-     */
+    /** Validates model policy and catalogue membership before settings are persisted. */
     validateModel(ref) {
       try {
         definition.models.assert(ref);
@@ -1109,11 +44,7 @@ export const createAgentKindService = <TState, TConfig extends object>(
     },
 
     listModels(caller) {
-      /**
-       * Which BYOK providers this caller has a key for. Read from the cookie rather than decrypted:
-       * the picker only needs to know whether a key is *present*, and decrypting to answer a listing
-       * request would put plaintext keys on a path that has no use for them.
-       */
+      // Listing only needs key presence; plaintext credentials never enter this path.
       const registered = readEncryptedAgentCredentials(caller.context.headers);
       const configured = BYOK_PROVIDER_IDS.filter(
         (providerId) => registered?.[providerId]
@@ -1125,6 +56,4 @@ export const createAgentKindService = <TState, TConfig extends object>(
       return Promise.resolve(definition.capabilities());
     },
   };
-
-  return service;
 };
